requirements.txt
fastapi
uvicorn
pytesseract
pillow
pdfplumber
python-multipart
python-docx
openpyxl
python-pptx
